# services/gateway-service/swagger-debug.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stock Analytics API",
    "version": "1.0.0",
    "description": "API documentation for the ExpressJS backend service",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/financials": {
      "get": {
        "summary": "Get financial data",
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "company",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Company ticker symbol"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "income-statement",
                "balance-sheet",
                "cash-flow"
              ]
            },
            "required": true,
            "description": "Type of financial statement"
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "annual",
                "quarterly"
              ]
            },
            "required": true,
            "description": "Reporting period"
          }
        ],
        "responses": {
          "200": {
            "description": "Financial data"
          }
        }
      }
    },
...
    "/stocks/{ticker}/financials": {
      "get": {
        "summary": "Get financial data for a stock",
        "tags": [
          "Stocks"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Stock ticker symbol"
          }
        ],
        "responses": {
          "200": {
            "description": "Financial data"
          }
        }
      }
    },
...
  },
...
}